+++ README.md
@@ -1,2 +1,17 @@
 # knock
+
 새로운 팀 노크의 react 레포지토리
+
+## git convention
+
+| 태그이름 | 설명                                               |
+| -------- | -------------------------------------------------- |
+| feat     | 새로운 기능 추가                                   |
+| fix      | 버그 수정                                          |
+| design   | css등 사용자 UI 수정                               |
+| style    | 코드 포맷 변경, 세미콜론 누락, 코드 수정 없는 경우 |
+| refactor | 프로덕션 코드 리팩토링                             |
+| comment  | 필요한 주석 추가 및 변경                           |
+| docs     | 문서 수정                                          |
+| rename   | 파일 또는 폴더명을 수정하거나 옮기는 작업          |
+| remove   | 파일을 삭제하는 작업만 수행                        |
