+++ src/app.jsx
@@ -1,10 +1,9 @@
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./css/style.css";
-import "./css/intro/intro.css";
-
+import "styles/global.css";
+import "styles/intro/intro.css";
 import { Suspense } from "react";
 import Routes from "./routes";
-import Header from "./components/header";
+import Header from "components/layout/header";
 import { BrowserRouter } from "react-router-dom";
 
 function App() {
