+++ src/lib/dummy_roommate.js
