+++ src/pages/home/guideline.jsx
