+++ src/pages/home/living_rule.jsx
