+++ src/pages/intro/no_home.jsx
